refactor move alignment business logic into useScriptureAlignment()
