+++ package.json
@@ -90,6 +90,6 @@
     "styled-components": "^5.2.1",
     "ts-loader": "^8.0.12",
     "typescript": "^4.1.3",
-    "word-aligner-rcl": "0.0.3-alpha.4"
+    "word-aligner-rcl": "0.0.3-alpha.5"
   }
 }
